+++ buildSrc/src/main/kotlin/Compiler.kt
@@ -2,6 +2,7 @@ object CompilerArguments {
     const val inlineClasses = "-XXLanguage:+InlineClasses"
     const val coroutines = "-Xopt-in=kotlinx.coroutines.ExperimentalCoroutinesApi"
     const val time = "-Xopt-in=kotlin.time.ExperimentalTime"
+    const val unstableDefault = "-Xopt-in=kotlinx.serialization.UnstableDefault"
 }
 
 object Jvm {